Redistributed instructions between "Device Setup" and "Starting the System"
How to power Wi-Fire boards section was after how to flash the board.
Also fixed an instance where the pages reference each other, and added a
note to the top of "Starting the System" to make it clear that all of
the "Device Setup" page needs to be completed first.